-Cleaning up warnings

+++ StereoKitC/systems/d3d.h
@@ -6,7 +6,7 @@
 #pragma comment(lib, "dxguid.lib") 
 #include <stdio.h>
 #define DX11ResType(res, name) res->SetPrivateData(WKPDID_D3DDebugObjectName, sizeof(name)-1, name);
-#define DX11ResName(res, type, name) { char tmp[256]; sprintf_s(tmp, 256, "%s|%s", type, name); res->SetPrivateData(WKPDID_D3DDebugObjectName, strlen(tmp), tmp); }
+#define DX11ResName(res, type, name) { char tmp[256]; sprintf_s(tmp, 256, "%s|%s", type, name); res->SetPrivateData(WKPDID_D3DDebugObjectName, (UINT)strlen(tmp), tmp); }
 #else
 #define DX11ResType(res, name)
 #define DX11ResName(res, type, name)
